refactor: big changes in server
Two instances of express was useless for the app, this changed. Routes
are scattered instead of being defined by using the Router function of
express, it was using the other express instances to do so, THIS DON'T
make sense when I took to look at the code again after 12 months.

Also, the same happens for the static files, it has two express.static
functions in the same folder, the css and js files are separated in
its own folder.